Benchmark (#24)
* Initiating AVX512 benchmarking...

* error fixing...

* resolved issues

* resolved issues after merge

* tried running the first AVX-benchmark

* bug fixing try 1..

* bug fixing try 2..

* bug fixing try 3..

* bug fixing try 4..

* bug fixing try 5..

* bug fixing try 6..

* somewhat functioning version..

* try fixing benchmark..

* trying to fix benchmark..

* Bugfixing RowStore-AVX

* configuring benchmark..

* trying to find something that works with pax

* Added Scalar Benchmark

* Fixed error..

* found another issue...

* Hotfix for row_store issue

* cleanup...

* added averaging to Benchmarks

* fixed type issue

* fixed iterator issue

* fixed ordering issue

* fixed normalization issue

* created exhaustive 32/64 incremental row benchmark

* fixing some errors..

* fixing some more errors..

* fixing some minor errors..

* fixing some major errors..

* fixing..

* fixing typecasting issue

* adding <compare> library for std::pair

* adding <compare> library for std::pair

* trying different string concatenation..

* adding const...

* fixing linking issue...

* fixing return issue...

* fixing pragma issue...

* fixing include issue...

* Fix gather 32 bit for column store

* Formatting

* Fix merge errors

* switch filter projection for row_store

* add filter 32bit benchmark

* Stop filtering after first mismatch in row store scalar

* Benchmark with all filters

* delete deletes..

* fix issues..

* fix memory leaks in Benchmark and PAXTable

* fix errors

* improve Benchmark

* remove memory leak...

* resolve issues

* resolve issues...

* Approve Pax destructor

* Free data in PaxTable.queryCount

* Added vector delete for scalar filtering

* decrease row count for sane results

* Fixed another memory leak in column store

* fix filter range

* adjusted filter ranges and indices

* adjusted row numbers

* adjusted filters (no eq)

+++ src/include/BenchmarkResult.h
@@ -7,7 +7,7 @@ template <typename T>
 class BenchmarkResult {
    private:
     // result map containing the parameters of the benchmark as a tuple as key and a tuple of the benchmark result as value
-    std::map<std::tuple<int, unsigned, unsigned, T, T>, std::tuple<uint64_t, uint64_t, double>> result;
+    std::map<std::tuple<int, unsigned, unsigned, T, T>, std::tuple<double, double, double>> result;
 
    public:
     /// Constructor
@@ -22,10 +22,10 @@ class BenchmarkResult {
     /// \param resultRowCount number of rows in queried table
     /// \param resultCpuTime CPU cycles needed for query
     /// \param resultRealTime time needed for query in ms
-    void addBenchmark(int tableStoreId, unsigned rowCount, unsigned columnCount, T lowerBound, T upperBound, uint64_t resultRowCount,
-                      uint64_t resultCpuTime, double resultRealTime) {
+    void addBenchmark(int tableStoreId, unsigned rowCount, unsigned columnCount, T lowerBound, T upperBound, double resultRowCount,
+                      double resultCpuTime, double resultRealTime) {
         std::tuple<int, unsigned, unsigned, T, T> key = {tableStoreId, rowCount, columnCount, lowerBound, upperBound};
-        std::tuple<uint64_t, uint64_t, double> value = {resultRowCount, resultCpuTime, resultRealTime};
+        std::tuple<double, double, double> value = {resultRowCount, resultCpuTime, resultRealTime};
 
         result.insert(std::pair{key, value});
     }
@@ -38,7 +38,7 @@ class BenchmarkResult {
         for (auto const& entry : result) {
             // separate key and value of map
             std::tuple<int, unsigned, unsigned, T, T> key = entry.first;
-            std::tuple<uint64_t, uint64_t, double> value = entry.second;
+            std::tuple<double, double, double> value = entry.second;
 
             // print key values
             std::cout << "Query parameters: " << std::endl;
@@ -69,7 +69,7 @@ class BenchmarkResult {
         for (auto const& entry : result) {
             // separate key and value of map
             std::tuple<int, unsigned, unsigned, T, T> key = entry.first;
-            std::tuple<uint64_t, uint64_t, double> value = entry.second;
+            std::tuple<double, double, double> value = entry.second;
 
             // add values to file
             file << std::get<0>(key) << "," << std::get<1>(key) << "," << std::get<2>(key) << "," << std::get<3>(key) << ","
